requirements.txt
fastmcp>=2.12.0
uvicorn>=0.35.0
python-dotenv>=1.0.0
requests>=2.31.0
pydexcom>=0.4.1
